requirements.txt
click>=8.1.7
pyyaml>=6.0.1
requests>=2.31.0
rich>=13.7.0
pydantic>=2.5.0
aiosqlite>=0.19.0

